Update README.md
Writing Your First Program

Develop a program that converts from kilometers to miles.

Read in a number that represents the number of kilometers traveled.

Convert this number to miles and then output it to the console.

1 kilometer = 0.621 miles.

To get started, make a copy of your solution from Lab 1.2 and paste it in the code editor panel. Modify this file to implement your solution to convert kilometers to miles.

Try running your program. If you get syntax errors, try to fix them and re-compile until your program is free of syntax errors. Is your output what you expect from the input you gave? If not, try to find and correct the logic error and run the program again. Continue this process until you have a program that produces the correct result.